Move VeryLargePart from Tests.cs to new file LargeFileTests.cs, mark as outerloop and disable parallelization.
