+++ src/libraries/System.IO.Packaging/tests/System.IO.Packaging.Tests.csproj
@@ -3,7 +3,10 @@
     <TargetFrameworks>$(NetCoreAppCurrent);$(NetFrameworkCurrent)</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="LargeFilesTests.cs" />
     <Compile Include="Tests.cs" />
+    <Compile Include="XunitAssemblyAttributes.cs" />
+    <Compile Include="$(CommonTestPath)TestUtilities\System\DisableParallelization.cs" Link="Common\TestUtilities\System\DisableParallelization.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFrameworkIdentifier)' != '.NETFramework'">
     <Compile Include="ReflectionTests.cs" />
